+++ src/modules/probabilistic-argumentation/GraphEditor.vue
@@ -36,6 +36,8 @@ import {
   type HistoryState,
   LinkType,
   type NodeId,
+  QUICK_EXPORT_KEY,
+  type QuickExport,
   type SelectionAction,
 } from '@/modules/common/graph-editor/graphEditor'
 import GraphEditor from '@/modules/common/graph-editor/GraphEditor.vue'
@@ -375,6 +377,10 @@ watch(evaluationHostOpen, (isOpen) => {
   if (!isOpen && layoutMode.value === 'compact') nodeWeights.value = new Map()
 })
 
+provide(QUICK_EXPORT_KEY, {
+  configs: availableExports as unknown as QuickExport['configs'],
+  getInput: () => state.current.content,
+})
 provide(TOOLTIP_REGISTRY_KEY, {
   ...abstractArgumentationGlossary,
   ...probabilisticArgumentationGlossary,
@@ -472,6 +478,7 @@ function onPopupKeydown(event: KeyboardEvent) {
       @redo="emit('redo')"
       @save="emit('save')"
       @share="emit('share')"
+      @export-file="emit('export', $event)"
       v-model:evaluation-open="evaluationHostOpen"
       @open-extension-window="addEvaluationInstance()"
     >